+++ tests/test_datetime.rs
@@ -2,7 +2,7 @@
 
 use pyo3::ffi::*;
 use pyo3::prelude::*;
-use pyo3::types::{IntoPyDict, PyAny};
+use pyo3::types::IntoPyDict;
 
 #[allow(clippy::trivially_copy_pass_by_ref)]
 fn _get_subclasses<'p>(